Added some utility for instruction decoding
dispatch.h: Added some functions to inline returning a constant in the dispatch with a function of the same name as the returned status

common.h: changed the extract[Register] function to static_cast the value into a register. Also added a function for rotating an immediate.

mask.h: added a strip function to strip off bits outside of the desired range for bit range. Probably gonna move it into the base mask struct

status.h + dispatch.h: added a new status to skip the current "branch" of decoding and return a status::nomatch to the "parent" dispatcher